SqlWriter/requirements.txt
kafka-python
pymysql
cryptography
simplejson
credittomodels
google
protobuf
numpy_financial
